+++ examples/resources/glesys_loadbalancer/import.sh
@@ -1,0 +1,2 @@
+# glesys_loadbalancer Import
+$ terraform import glesys_loadbalancer.mylb lb123456