+++ examples/resources/glesys_loadbalancer_backend/import.sh
@@ -1,0 +1,2 @@
+# glesys_loadbalancer_backend Import
+$ terraform import glesys_loadbalancer_backend.be lb123456,http-backend-name