+++ glesys/utils.go
@@ -1,0 +1,20 @@
+package glesys
+
+import (
+	"context"
+	"strings"
+
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+)
+
+func resourceGlesysLoadbalancerBeFeImport(ctx context.Context, d *schema.ResourceData, m interface{}) ([]*schema.ResourceData, error) {
+	if strings.Contains(d.Id(), ",") {
+		s := strings.Split(d.Id(), ",")
+
+		d.SetId(s[1])
+		d.Set("name", s[1])
+		d.Set("loadbalancerid", s[0])
+	}
+
+	return []*schema.ResourceData{d}, nil
+}